Make loaders discoverable
This makes loaders for a Ginga viewer discoverable via entry points.
This will make it possible to write 3rd-party loaders that are not
bundled with Ginga.

Loaders are consolidated in ginga.util.io. Legacy stubs with warning
on import are left in ginga.util.io_rgb and ginga.util.io_fits.
