+++ ginga/util/colorramp.py
@@ -40,7 +40,7 @@ def rgbarr_to_colormap(rgb_data_np):
 
 def png_to_colormap(png_file):
 
-    from ginga.util import io_rgb
+    from ginga.util.io import io_rgb
     img = io_rgb.load_file(png_file)
     data = img.get_data()
