+++ setup.cfg
@@ -98,6 +98,13 @@ ginga_modes =
     camera = ginga.modes.camera:CameraMode
     naxis = ginga.modes.naxis:NaxisMode
 
+ginga_loaders =
+    pillow = ginga.util.io.io_rgb:PillowFileHandler
+    opencv = ginga.util.io.io_rgb:OpenCvFileHandler
+    astropy = ginga.util.io.io_fits:AstropyFitsFileHandler
+    fitsio = ginga.util.io.io_fits:FitsioFileHandler
+    asdf = ginga.util.io.io_asdf:ASDFFileHandler
+
 [flake8]
 # Ignoring these for now:
 # E129: visually indented line with same indent as next logical line
